# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MCEvents API",
    "version": "1.0.0",
    "description": "The official API for the MCEvents Minecraft server.\n\n**NOTE: This API uses your IP for rate limiting.** These are not stored and is handled by the middleware package [express-rate-limit](https://www.npmjs.com/package/express-rate-limit).\n\nThe current rate is 100 requests per 15 minutes. **This may change in the future, as this is an early release.**\n\n[Source code](https://github.com/mcevents/MCEventsAPI)",
    "license": {
      "name": "GPL 3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "contact": {
      "name": "ggtylerr",
      "url": "https://ggtylerr.dev",
      "email": "ggtylerr_contact@protonmail.com"
    }
  },
...
  "paths": {
...
    "/v1/lb/wins": {
      "get": {
        "summary": "Gets the top winners.",
        "tags": ["Leaderboard"],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "uuid": {
                                "type": "string",
                                "example": "d2c95f31-40cb-4035-b13c-fc7e00aa8940"
                              },
                              "user": {
                                "type": "string",
                                "example": "Sourblock"
                              },
                              "wins": {
                                "type": "integer",
                                "example": 0
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer",
                          "example": 500
                        },
                        "message": {
                          "type": "string",
                          "example": "DB error."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/lb/kills": {
      "get": {
        "summary": "Gets the top murderers.",
        "tags": ["Leaderboard"],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "uuid": {
                                "type": "string",
                                "example": "d2c95f31-40cb-4035-b13c-fc7e00aa8940"
                              },
                              "user": {
                                "type": "string",
                                "example": "Sourblock"
                              },
                              "kills": {
                                "type": "integer",
                                "example": 0
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "res": {
                      "type": "object"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer",
                          "example": 500
                        },
                        "message": {
                          "type": "string",
                          "example": "DB error."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}